+++ pkg/server/config/config_test.go
@@ -401,6 +401,53 @@ func TestVerifyConfig(t *testing.T) {
 		})
 	})
 
+	t.Run("list_objects_iterator_cache", func(t *testing.T) {
+		t.Run("enable_but_ttl_zero", func(t *testing.T) {
+			cfg := DefaultConfig()
+			cfg.ListObjectsIteratorCache.Enabled = true
+			cfg.ListObjectsIteratorCache.TTL = 0
+			cfg.ListObjectsIteratorCache.MaxResults = 1000
+			err := cfg.Verify()
+			require.Error(t, err)
+		})
+
+		t.Run("enable_but_ttl_negative", func(t *testing.T) {
+			cfg := DefaultConfig()
+			cfg.ListObjectsIteratorCache.Enabled = true
+			cfg.ListObjectsIteratorCache.TTL = -2 * time.Second
+			cfg.ListObjectsIteratorCache.MaxResults = 1000
+			err := cfg.Verify()
+			require.Error(t, err)
+		})
+
+		t.Run("enable_but_max_results_zero", func(t *testing.T) {
+			cfg := DefaultConfig()
+			cfg.ListObjectsIteratorCache.Enabled = true
+			cfg.ListObjectsIteratorCache.TTL = 2 * time.Second
+			cfg.ListObjectsIteratorCache.MaxResults = 0
+			err := cfg.Verify()
+			require.Error(t, err)
+		})
+
+		t.Run("disable_but_ttl_and_max_results_zero", func(t *testing.T) {
+			cfg := DefaultConfig()
+			cfg.ListObjectsIteratorCache.Enabled = false
+			cfg.ListObjectsIteratorCache.TTL = 0
+			cfg.ListObjectsIteratorCache.MaxResults = 0
+			err := cfg.Verify()
+			require.NoError(t, err)
+		})
+
+		t.Run("enable_and_ttl_and_max_results_positive", func(t *testing.T) {
+			cfg := DefaultConfig()
+			cfg.ListObjectsIteratorCache.Enabled = true
+			cfg.ListObjectsIteratorCache.TTL = 10 * time.Second
+			cfg.ListObjectsIteratorCache.MaxResults = 10000
+			err := cfg.Verify()
+			require.NoError(t, err)
+		})
+	})
+
 	t.Run("cache_controller", func(t *testing.T) {
 		t.Run("enable_but_ttl_zero", func(t *testing.T) {
 			cfg := DefaultConfig()