+++ pkg/server/server_test.go
@@ -2123,9 +2123,9 @@ func TestServerCheckCache(t *testing.T) {
 		)
 		t.Cleanup(s.Close)
 
-		require.NotNil(t, s.sharedCheckResources.CheckCache)
+		require.NotNil(t, s.sharedDatastoreResources.CheckCache)
 		require.True(t, s.cacheSettings.ShouldCacheCheckQueries())
-		require.True(t, s.cacheSettings.ShouldCacheIterators())
+		require.True(t, s.cacheSettings.ShouldCacheCheckIterators())
 	})
 
 	t.Run("query_cache_disabled_iterator_cache_enabled", func(t *testing.T) {
@@ -2139,7 +2139,7 @@ func TestServerCheckCache(t *testing.T) {
 		)
 		t.Cleanup(s.Close)
 
-		require.NotNil(t, s.sharedCheckResources.CheckCache)
+		require.NotNil(t, s.sharedDatastoreResources.CheckCache)
 	})
 
 	t.Run("query_cache_enabled_iterator_cache_disabled", func(t *testing.T) {
@@ -2153,7 +2153,7 @@ func TestServerCheckCache(t *testing.T) {
 		)
 		t.Cleanup(s.Close)
 
-		require.NotNil(t, s.sharedCheckResources.CheckCache)
+		require.NotNil(t, s.sharedDatastoreResources.CheckCache)
 	})
 
 	t.Run("query_cache_disabled_iterator_cache_disabled", func(t *testing.T) {
@@ -2167,7 +2167,51 @@ func TestServerCheckCache(t *testing.T) {
 		)
 		t.Cleanup(s.Close)
 
-		require.Nil(t, s.sharedCheckResources.CheckCache)
+		require.Nil(t, s.sharedDatastoreResources.CheckCache)
+	})
+}
+
+func TestServerListObjectsCache(t *testing.T) {
+	t.Cleanup(func() {
+		goleak.VerifyNone(t)
+	})
+
+	t.Run("list_objects_cache_enabled", func(t *testing.T) {
+		s := MustNewServerWithOpts(
+			WithDatastore(memory.New()),
+			WithListObjectsIteratorCacheEnabled(true),
+		)
+		t.Cleanup(s.Close)
+
+		require.NotNil(t, s.sharedDatastoreResources.CheckCache)
+		require.True(t, s.cacheSettings.ShouldCacheListObjectsIterators())
+		require.False(t, s.cacheSettings.ShouldCacheCheckQueries())
+		require.False(t, s.cacheSettings.ShouldCacheCheckIterators())
+	})
+
+	t.Run("all_caches_enabled", func(t *testing.T) {
+		s := MustNewServerWithOpts(
+			WithDatastore(memory.New()),
+			WithCheckQueryCacheEnabled(true),
+			WithCheckQueryCacheTTL(1*time.Minute),
+			WithCheckCacheLimit(10),
+			WithCheckIteratorCacheEnabled(true),
+			WithCheckIteratorCacheMaxResults(10),
+			WithListObjectsIteratorCacheEnabled(true),
+		)
+		t.Cleanup(s.Close)
+
+		require.NotNil(t, s.sharedDatastoreResources.CheckCache)
+		require.True(t, s.cacheSettings.ShouldCacheListObjectsIterators())
+		require.True(t, s.cacheSettings.ShouldCacheCheckQueries())
+		require.True(t, s.cacheSettings.ShouldCacheCheckIterators())
+	})
+
+	t.Run("no_caches_enabled", func(t *testing.T) {
+		s := MustNewServerWithOpts(WithDatastore(memory.New()))
+		t.Cleanup(s.Close)
+
+		require.Nil(t, s.sharedDatastoreResources.CheckCache)
 	})
 }
 
@@ -2191,8 +2235,8 @@ func TestCheckWithCachedControllerEnabled(t *testing.T) {
 			s.Close()
 		})
 
-		require.NotNil(t, s.sharedCheckResources.CacheController)
-		_, ok := s.sharedCheckResources.CacheController.(*cachecontroller.NoopCacheController)
+		require.NotNil(t, s.sharedDatastoreResources.CacheController)
+		_, ok := s.sharedDatastoreResources.CacheController.(*cachecontroller.NoopCacheController)
 		require.True(t, ok)
 	})
 
@@ -2212,7 +2256,7 @@ func TestCheckWithCachedControllerEnabled(t *testing.T) {
 			s.Close()
 		})
 
-		require.NotNil(t, s.sharedCheckResources.CacheController)
+		require.NotNil(t, s.sharedDatastoreResources.CacheController)
 	})
 
 	t.Run("cache_controller_is_not_nil_if_check_iterator_cache_enabled", func(t *testing.T) {
@@ -2231,7 +2275,7 @@ func TestCheckWithCachedControllerEnabled(t *testing.T) {
 			s.Close()
 		})
 
-		require.NotNil(t, s.sharedCheckResources.CacheController)
+		require.NotNil(t, s.sharedDatastoreResources.CacheController)
 	})
 }